output: write individual tree trajectories to feather

+++ iLand/Extensions/SpanExtensions.cs
@@ -4,11 +4,11 @@ namespace iLand.Extensions
 {
     internal static class SpanExtensions
     {
-        public static void FillIncrementing(this Span<int> span, int initialValue)
+        public static void FillIncrementing(this Span<Int16> span, int initialValue)
         {
             for (int index = 0; index < span.Length; ++index)
             {
-                span[index] = initialValue + index;
+                span[index] = (Int16)(initialValue + index);
             }
         }
     }
